pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spotify-wav-downloader"
version = "0.2.0"
description = "Download .wav files from Spotify playlists in the highest quality lossless audio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "spotipy>=2.23.0",
    "yt-dlp>=2024.1.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "mutagen>=1.47.0",
]

[project.scripts]
spotify-wav-dl = "spotify_wav_dl.cli:main"

[tool.setuptools.packages.find]
include = ["spotify_wav_dl*"]
